Refactored data saving, roll display
Creature data saving was changed so there is a link between the creature type in the creature list and a creature of the same type in initiative.
Changes to the statblock of the creature type will now reflect in the creatures in initiative.
The numerical creature identifier in combat will now also display when it makes a roll (e.g. "Kobold 5" instead of "Kobold")