Added saving of roll data
Added saving of roll data: rolls are now saved in the database. A history of previous rolls is loaded and displayed on the initiative page.
Adjusted UI of roll display so it does not shrink when the creature Ui is expanded, and so three rolls can be viewed at any time.